pyproject.toml
[project]
name = "matrx-ai"
version = "0.1.0"
description = "Matrx AI — core FastAPI backend for AI integrations"
requires-python = ">=3.13"

dependencies = [
    # --- Web framework & server ---
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "uvloop>=0.21.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sse-starlette>=2.2.0",
    "httpx>=0.28.0",
    "orjson>=3.10.0",
    "python-dotenv>=1.0.0",

    # --- AI providers ---
    "anthropic",
    "openai",
    "google-genai",
    "cerebras-cloud-sdk",
    "groq",
    "together",
    "xai-sdk",

    # --- Supabase & database ---
    "supabase",

    # --- Matrx ecosystem ---
    "matrx-utils>=1.0.4",
    "matrx-orm>=1.5.3",

    # --- Utilities used by ai module ---
    "rich",
    "tiktoken",
    "aiofiles",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pyright>=1.1.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pyright>=1.1.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "ASYNC"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "standard"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
